Split up routes into separate files. * booksAPI.js - /api/    * All API functionality. * login.js - /login/    * Cookies and password checking. * basic.js - /    * Serve index.html.
