+++ routes/basic.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const path = require('path');
+const router = express.Router();
+
+router.get('/', (req, res, next) => {
+  res.sendFile(path.join(
+    __dirname, '..', '/public/', 'index.html'));
+});
+
+module.exports = router;
